+++ sys/shell/commands/sc_gnrc_ipv6_nib.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2017 Freie Universität Berlin
+ * Copyright (C) 2019 FZI Forschungszentrum Informatik
  *
  * This file is subject to the terms and conditions of the GNU Lesser
  * General Public License v2.1. See the file LICENSE in the top level
@@ -261,7 +262,7 @@ static int _nib_route(int argc, char **argv)
             return 1;
         }
         if (argc > 6) {
-            ltime = (uint16_t)atoi(argv[6]);
+            ltime = atoi(argv[6]);
         }
         gnrc_ipv6_nib_ft_add(&pfx, pfx_len, &next_hop, iface, ltime);
     }
