+++ tests/periph_pwm/main.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2014-2015 Freie Universität Berlin
+ * Copyright (C) 2019 FZI Forschungszentrum Informatik
  *
  * This file is subject to the terms and conditions of the GNU Lesser
  * General Public License v2.1. See the file LICENSE in the top level
@@ -46,7 +47,7 @@ static uint32_t initiated;
 
 static unsigned _get_dev(const char *dev_str)
 {
-    unsigned dev = (unsigned)atoi(dev_str);
+    unsigned dev = atoi(dev_str);
     if (dev >= PWM_NUMOF) {
         printf("Error: device PWM_DEV(%u) is unknown\n", dev);
         return UINT_MAX;
@@ -123,7 +124,7 @@ static int _set(int argc, char**argv)
         return 1;
     }
 
-    uint8_t chan = (uint8_t)atoi(argv[2]);
+    uint8_t chan = atoi(argv[2]);
     if (chan >= pwm_channels(PWM_DEV(dev))) {
         printf("Error: channel %d is unknown.\n", chan);
         return 1;
