+++ tests/ps_schedstatistics/main.c
@@ -1,6 +1,7 @@
 /*
  * Copyright (C) 2017 OTA keys S.A.
  *               2017 HAW Hamburg
+ * Copyright (C) 2019 FZI Forschungszentrum Informatik
  *
  * This file is subject to the terms and conditions of the GNU Lesser
  * General Public License v2.1. See the file LICENSE in the top level
@@ -34,9 +35,9 @@ static kernel_pid_t pids[NB_THREADS];
 
 static void *_thread_fn(void *arg)
 {
-    int next = ((int)arg + 1) % NB_THREADS;
+    int next = ((uintptr_t)arg + 1) % NB_THREADS;
 
-    printf("Creating thread #%d, next=%d\n", (int)arg, next);
+    printf("Creating thread #%"PRIdPTR", next=%d\n", (uintptr_t) arg, next);
 
     while (1) {
         msg_t m1, m2;
@@ -54,7 +55,7 @@ static void *_thread_fn(void *arg)
 
 int main(void)
 {
-    for (unsigned i = 0; i < NB_THREADS; ++i) {
+    for (uintptr_t i = 0; i < NB_THREADS; ++i) {
         pids[i] = thread_create(stacks[i], sizeof(stacks[i]),
                                 THREAD_PRIORITY_MAIN - 1,
                                 THREAD_CREATE_STACKTEST,
